refactor: common initOptions
