+++ lib/browser/client.js
@@ -1,11 +1,9 @@
-
 const debug = require('debug')('ali-oss');
 const crypto = require('crypto');
 const path = require('path');
 const copy = require('copy-to');
 const mime = require('mime');
 const xml = require('xml2js');
-const ms = require('humanize-ms');
 const AgentKeepalive = require('agentkeepalive');
 const merge = require('merge-descriptors');
 const urlutil = require('url');
@@ -18,6 +16,7 @@ const dateFormat = require('dateformat');
 const bowser = require('bowser');
 const signUtils = require('../common/signUtils');
 const utils = require('../common/utils');
+const _initOptions = require('../common/client/initOptions');
 
 const globalHttpAgent = new AgentKeepalive();
 
@@ -25,35 +24,10 @@ function getHeader(headers, name) {
   return headers[name] || headers[name.toLowerCase()];
 }
 
-function setEndpoint(endpoint, secure) {
-  let url = urlutil.parse(endpoint);
-
-  if (!url.protocol) {
-    const protocol = secure ? 'https://' : 'http://';
-    url = urlutil.parse(protocol + endpoint);
-  }
-
-  if (url.protocol !== 'http:' && url.protocol !== 'https:') {
-    throw new Error('Endpoint protocol must be http or https.');
-  }
-
-  return url;
-}
-
-function setRegion(region, internal, secure) {
-  const protocol = secure ? 'https://' : 'http://';
-  let suffix = internal ? '-internal.aliyuncs.com' : '.aliyuncs.com';
-  const prefix = 'vpc100-oss-cn-';
-  // aliyun VPC region: https://help.aliyun.com/knowledge_detail/38740.html
-  if (region.substr(0, prefix.length) === prefix) {
-    suffix = '.aliyuncs.com';
-  }
-
-  return urlutil.parse(protocol + region + suffix);
-}
-
 // check local web protocol,if https secure default set true , if http secure default set false
 function isHttpsWebProtocol() {
+  // for web worker not use window.location.
+  // eslint-disable-next-line no-restricted-globals
   return location && location.protocol === 'https:';
 }
 
@@ -93,47 +67,13 @@ Client.initOptions = function initOptions(options) {
   if (!options.stsToken) {
     console.warn('Please use STS Token for safety, see more details at https://help.aliyun.com/document_detail/32077.html');
   }
+  const opts = Object.assign({
+    secure: isHttpsWebProtocol(),
+    // for browser compatibility disable fetch.
+    useFetch: false
+  }, options);
 
-  if (!options
-    || !options.accessKeyId
-    || !options.accessKeySecret) {
-    throw new Error('require accessKeyId, accessKeySecret');
-  }
-
-  const isHttpsProtocol = isHttpsWebProtocol();
-  const opts = {
-    region: 'oss-cn-hangzhou',
-    internal: false,
-    secure: isHttpsProtocol,
-    bucket: null,
-    endpoint: null,
-    cname: false,
-    isRequestPay: false,
-    useFetch: true
-  };
-
-  Object.keys(options).forEach((key) => {
-    if (options[key] !== undefined) {
-      opts[key] = options[key];
-    }
-  });
-  opts.accessKeyId = opts.accessKeyId.trim();
-  opts.accessKeySecret = opts.accessKeySecret.trim();
-
-  if (opts.timeout) {
-    opts.timeout = ms(opts.timeout);
-  }
-
-  if (opts.endpoint) {
-    opts.endpoint = setEndpoint(opts.endpoint, opts.secure);
-  } else if (opts.region) {
-    opts.endpoint = setRegion(opts.region, opts.internal, opts.secure);
-  } else {
-    throw new Error('require options.endpoint or options.region');
-  }
-
-  opts.inited = true;
-  return opts;
+  return _initOptions(opts);
 };
