+++ lib/bucket.js
@@ -1,7 +1,7 @@
 
 
 const assert = require('assert');
-const utils = require('./common/utils/checkBucketname');
+const utils = require('./common/utils/checkBucketName');
 
 const proto = exports;
 
@@ -18,7 +18,7 @@ function toArray(obj) {
 }
 
 /**
- * check Bucket name
+ * check Bucket Name
  */
 
 proto._checkBucketName = function _checkBucketName(name) {
