+++ lib/client.js
@@ -6,7 +6,6 @@ const path = require('path');
 const copy = require('copy-to');
 const mime = require('mime');
 const xml = require('xml2js');
-const ms = require('humanize-ms');
 const AgentKeepalive = require('agentkeepalive');
 const HttpsAgentKeepalive = require('agentkeepalive').HttpsAgent;
 const merge = require('merge-descriptors');
@@ -20,6 +19,7 @@ const dateFormat = require('dateformat');
 const bowser = require('bowser');
 const signUtils = require('./common/signUtils');
 const utils = require('./common/utils');
+const _initOptions = require('./common/client/initOptions');
 
 const globalHttpAgent = new AgentKeepalive();
 const globalHttpsAgent = new HttpsAgentKeepalive();
@@ -28,31 +28,6 @@ function getHeader(headers, name) {
   return headers[name] || headers[name.toLowerCase()];
 }
 
-function setEndpoint(endpoint) {
-  let url = urlutil.parse(endpoint);
-
-  if (!url.protocol) {
-    url = urlutil.parse(`http://${endpoint}`);
-  }
-
-  if (url.protocol !== 'http:' && url.protocol !== 'https:') {
-    throw new Error('Endpoint protocol must be http or https.');
-  }
-
-  return url;
-}
-
-function setRegion(region, internal, secure) {
-  const protocol = secure ? 'https://' : 'http://';
-  let suffix = internal ? '-internal.aliyuncs.com' : '.aliyuncs.com';
-  const prefix = 'vpc100-oss-cn-';
-  // aliyun VPC region: https://help.aliyun.com/knowledge_detail/38740.html
-  if (region.substr(0, prefix.length) === prefix) {
-    suffix = '.aliyuncs.com';
-  }
-
-  return urlutil.parse(protocol + region + suffix);
-}
 
 function Client(options, ctx) {
   if (!(this instanceof Client)) {
@@ -84,43 +59,7 @@ function Client(options, ctx) {
 module.exports = Client;
 
 Client.initOptions = function initOptions(options) {
-  if (!options
-    || !options.accessKeyId
-    || !options.accessKeySecret) {
-    throw new Error('require accessKeyId, accessKeySecret');
-  }
-
-  const opts = {
-    region: 'oss-cn-hangzhou',
-    internal: false,
-    secure: false,
-    timeout: 60000, // 60s
-    bucket: null,
-    endpoint: null,
-    cname: false,
-    isRequestPay: false
-  };
-
-  Object.keys(options).forEach((key) => {
-    if (options[key] !== undefined) {
-      opts[key] = options[key];
-    }
-  });
-  opts.accessKeyId = opts.accessKeyId.trim();
-  opts.accessKeySecret = opts.accessKeySecret.trim();
-
-  opts.timeout = ms(opts.timeout);
-
-  if (opts.endpoint) {
-    opts.endpoint = setEndpoint(opts.endpoint);
-  } else if (opts.region) {
-    opts.endpoint = setRegion(opts.region, opts.internal, opts.secure);
-  } else {
-    throw new Error('require options.endpoint or options.region');
-  }
-
-  opts.inited = true;
-  return opts;
+  return _initOptions(options);
 };
 
 /**
