+++ lib/common/utils/checkBucketName.js
@@ -1,5 +1,5 @@
 /**
- * check Bucket name
+ * check Bucket Name
  */
 
 exports._checkBucketName = function (name) {
